refactor discount

+++ src/main/java/com/epam/www/domain/Price.java
@@ -1,9 +1,11 @@
 package com.epam.www.domain;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Zoltan_Biro on 5/16/2017.
  */
 public interface Price {
 
-    long getSumPrice();
+    BigDecimal getSumPrice();
 }
