+++ src/main/java/com/epam/www/service/PricingService.java
@@ -1,9 +1,9 @@
 package com.epam.www.service;
 
 import com.epam.www.domain.DiscountEnums;
-import com.epam.www.domain.BasePrice;
 import com.epam.www.domain.Price;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -11,6 +11,6 @@ import java.util.List;
  */
 public interface PricingService {
 
-    Price getPrice(int normalSeats, int vipSeats, long price, List<DiscountEnums> discounts);
+    Price getPrice(int normalSeats, int vipSeats, BigDecimal price, List<DiscountEnums> discounts);
 }
